# rest-api/src/main/resources/api.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Hexa API
  description: Exemple d'une API pizza simple sur architecture hexagonale.
tags:
  - name: hexa
paths:
  /pizzas:
    get:
      operationId: getPizzas
      description: Récupère la liste des pizzas
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
          required: false
        - in: query
          name: size
          schema:
            type: integer
            default: 20
          required: false
      tags:
        - pizza
      responses:
        '200':
          description: Succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/pizzaDTO'
    post:
      operationId: createPizza
      description: Crée une nouvelle pizza
      tags:
        - pizza
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/nouvellePizzaDTO'
      responses:
        '200':
          description: Succès
        '400':
          description: Requête invalide
  /pizzas/{pizzaId}:
    get:
      operationId: getPizza
      description: Récupère une pizza par son identifiant
      parameters:
        - $ref: '#/components/parameters/pizzaId'
      tags:
        - pizza
      responses:
        '200':
          description: Succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pizzaDTO'
        '404':
          description: Pizza introuvable
    put:
      operationId: updatePizza
      description: Met à jour une pizza
      parameters:
        - $ref: '#/components/parameters/pizzaId'
      tags:
        - pizza
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/pizzaDTO'
      responses:
        '200':
          description: Succès
        '400':
          description: Requête invalide
    delete:
      operationId: deletePizza
      description: Supprime la pizza identifiée
      parameters:
        - $ref: '#/components/parameters/pizzaId'
      tags:
        - pizza
      responses:
        '200':
          description: Succès
        '404':
          description: Pizza introuvable
    patch:
      operationId: altererPizza
      description: Modifie la pizza sur certains aspects (cuisson, ...)
      parameters:
        - $ref: '#/components/parameters/pizzaId'
      tags:
        - pizza
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/pizzaAltereeDTO'
      responses:
        '204':
          description: Succès
components:
  schemas:
    nouvellePizzaDTO:
      type: object
      required:
        - taille
        - ingredients
      properties:
        taille:
          type: string
          enum:
            - ENFANT
            - MEDIUM
            - LARGE
            - TRES_GROS_MANGEUR
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/ingredientDTO"
    pizzaDTO:
      allOf:
        - $ref: "#/components/schemas/nouvellePizzaDTO"
        - type: object
          properties:
            id:
              type: integer
    pizzaAltereeDTO:
      type: object
      required:
        - cuire
      properties:
        cuire:
          type: boolean
    ingredientDTO:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          items:
            type: integer
  parameters:
    pizzaId:
      name: pizzaId
      in: path
      description: ID of a pizza
      required: true
      schema:
        type: long